feat: refactored the edit screen
